requirements.txt
# Enhanced PII Redactor Requirements
# Core dependencies for 99% accuracy PII detection system

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Azure AI Services
azure-ai-textanalytics>=5.3.0
azure-core>=1.24.0

# HTTP Requests
requests>=2.28.0

# Machine Learning (Optional but recommended for enhanced accuracy)
scikit-learn>=1.3.0
joblib>=1.3.0

# Web Framework
streamlit>=1.28.0

# Visualization
plotly>=5.15.0

# PDF Generation
reportlab>=4.0.0

# Environment Management
python-dotenv>=1.0.0

# Data Validation
jsonschema>=4.17.0

# Development and Testing (Optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Performance Monitoring
psutil>=5.9.0

# Additional utilities
tqdm>=4.65.0  # Progress bars
colorama>=0.4.6  # Colored terminal output
openpyxl>=3.1.2  # For Excel file support
